Bug fixes, added basic Unity class to handle events
About EventQueue.cs:

EventQueue.cs is a simple event manager where its events (classes that
implement Event as base class) will be called consecutively when the
last event finishes. This enables behavior where you can for example
launch multiple message boxes, or a time-based transition next to each
other and they will not overlap. Coupled with the Coffee Flow app to
control the flow of what methods get fired, this can prove to be handy
when using the tool for video game development.